[Feat] 댓글 API 로그인 인증 미들웨어 추가

+++ services/comment.js
@@ -1,4 +1,4 @@
-const { postModel, commentModel } = require('../models');
+const { postModel, commentModel, userModel } = require('../models');
 const errorCodes = require('../utils/errorCodes');
 
 /**
@@ -7,15 +7,21 @@ const errorCodes = require('../utils/errorCodes');
  * @function addComment
  * @param {Object} commentInfo comment 생성 시 필요한 정보
  * @param {String} commentInfo.postId 게시글 id
+ * @param {String} commentInfo.comUserId 유저 id
  * @param {String} commentInfo.comment 댓글 내용
  * @returns {Object} 생성된 comment 객체
  */
 const addComment = async (commentInfo) => {
-  const { postId } = commentInfo;
-  const post = await postModel.findPost(postId);
+  const { postId, comUserId } = commentInfo;
 
+  const post = await postModel.findPost(postId);
   if (!post) {
     throw new Error(errorCodes.canNotFindPost);
+  }
+
+  const user = await userModel.findUserById(comUserId);
+  if (!user) {
+    throw new Error(errorCodes.canNotFindUser);
   }
 
   // 해당 게시글에 대한 댓글 확인
@@ -39,21 +45,25 @@ const addComment = async (commentInfo) => {
  * @function addReply
  * @param {Object} replyInfo reply 생성 시 필요한 정보
  * @param {String} replyInfo.postId 게시글 id
+ * @param {String} replyInfo.comUserId 유저 id
  * @param {String} replyInfo.commentId 댓글 id
  * @param {String} replyInfo.comment 대댓글 내용
  * @returns {Object} 생성된 reply 객체
  */
 const addReply = async (replyInfo) => {
-  const { postId, commentId } = replyInfo;
+  const { postId, commentId, comUserId } = replyInfo;
 
   const post = await postModel.findPost(postId);
-
   if (!post) {
     throw new Error(errorCodes.canNotFindPost);
   }
 
-  const parentComment = await commentModel.findComment(commentId);
+  const user = await userModel.findUserById(comUserId);
+  if (!user) {
+    throw new Error(errorCodes.canNotFindUser);
+  }
 
+  const parentComment = await commentModel.findComment(commentId);
   if (!parentComment) {
     throw new Error(errorCodes.canNotFindComment);
   }
@@ -79,7 +89,6 @@ const addReply = async (replyInfo) => {
  */
 const getComments = async (postId) => {
   const post = await postModel.findPost(postId);
-
   if (!post) {
     throw new Error(errorCodes.canNotFindPost);
   }
@@ -94,21 +103,31 @@ const getComments = async (postId) => {
  * @function setComment
  * @param {Object} updateInfo 수정 정보
  * @param {String} updateInfo.postId 게시글 id
+ * @param {String} updateInfo.comUserId 유저 id
  * @param {String} updateInfo.commentId 댓글 id
  * @param {String} updateInfo.comment 수정 내용
  * @returns {Array} 댓글 내용 수정 결과
  */
 const setComment = async (updateInfo) => {
-  const { postId, commentId } = updateInfo;
+  const { postId, commentId, comUserId } = updateInfo;
 
   const post = await postModel.findPost(postId);
   if (!post) {
     throw new Error(errorCodes.canNotFindPost);
   }
 
+  const user = await userModel.findUserById(comUserId);
+  if (!user) {
+    throw new Error(errorCodes.canNotFindUser);
+  }
+
   const comment = await commentModel.findComment(commentId);
   if (!comment) {
     throw new Error(errorCodes.canNotFindComment);
+  }
+
+  if (user.id !== comment.comUserId) {
+    throw new Error(errorCodes.notMatchedUser);
   }
 
   const reuslt = await commentModel.updateComment(updateInfo);
@@ -122,20 +141,30 @@ const setComment = async (updateInfo) => {
  * @function deleteComment
  * @param {Object} deleteInfo 삭제 정보
  * @param {String} deleteInfo.postId 게시글 id
+ * @param {String} deleteInfo.comUserId 유저 id
  * @param {String} deleteInfo.commentId 댓글 id
  * @returns {Number} 댓글 삭제 결과
  */
 const deleteComment = async (deleteInfo) => {
-  const { postId, commentId } = deleteInfo;
+  const { postId, commentId, comUserId } = deleteInfo;
 
   const post = await postModel.findPost(postId);
   if (!post) {
     throw new Error(errorCodes.canNotFindPost);
   }
 
+  const user = await userModel.findUserById(comUserId);
+  if (!user) {
+    throw new Error(errorCodes.canNotFindUser);
+  }
+
   const comment = await commentModel.findComment(commentId);
   if (!comment) {
     throw new Error(errorCodes.canNotFindComment);
+  }
+
+  if (user.id !== comment.comUserId) {
+    throw new Error(errorCodes.notMatchedUser);
   }
 
   const reuslt = await commentModel.destroyComment(commentId);
